refactor. UI로직과 비즈니스 로직을 분리
 - RacingCarGame에서 사용되는 UI로직을 Application으로 이동
