--- src/main/java/domain/Application.java
@@ -1,8 +1,0 @@
-package domain;
-
-public class Application {
-    public static void main(String[] args) {
-        RacingCarGame game = new RacingCarGame();
-        game.run();
-    }
-}
