+++ src/main/java/racingcargame/view/InputView.java
@@ -1,4 +1,4 @@
-package view;
+package racingcargame.view;
 
 import java.util.Scanner;
